// types/interface/test.ts
export interface TestQuestion {
  ma_cau_hoi: number;             // question_id
  ma_bai_kiem_tra: number;           // test_id
  // ma_phan_kiem_tra: number;          // section_id
  loai: "trac_nghiem" | "tu_luan" | "dung_sai" | "tra_loi_ngan"; // type
  cau_hoi: string;                   // question
  lua_chon: string[];               // options
  dap_an_dung?: string | number;     // correctAnswer
  diem: number;                      // points
  giai_thich?: string;               // explanation
  ngay_tao: string;                  // createdAt
  ngay_cap_nhat: string;             // updatedAt
}